add dynamic memory function.

+++ kernel/type.h
@@ -1,10 +1,10 @@
 #ifndef __TYPE_H__
 #define __TYPE_H__
 
-typedef unsigned char   uint8;
-typedef unsigned short  uint16;
-typedef unsigned int    uint32;
-typedef unsigned long   uint64;
-typedef uint8           byte;
+typedef      unsigned   char   uint8;
+typedef      unsigned   short  uint16;
+typedef      unsigned   int    uint32;
+typedef      unsigned   long   uint64;
+typedef long unsigned   int    size_t;
 
 #endif
